Footer mobile version and loading animation
